Challenge_1b/requirements.txt
pdfminer.six==20220506
rank-bm25==0.2.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.31.0
numpy==1.24.4
scipy==1.10.1
regex==2023.6.3
tqdm==4.67.1